+++ IntuneBackupAndRestore/Public/Invoke-IntuneRestoreAppProtectionPolicy.ps1
@@ -22,10 +22,8 @@ function Invoke-IntuneRestoreAppProtectionPolicy {
         [string]$ApiVersion = "Beta"
     )
 
-    #Connect to MS-Graph if required
-    if ($null -eq (Get-MgContext)) {
-        connect-mggraph -scopes "DeviceManagementApps.ReadWrite.All, DeviceManagementConfiguration.ReadWrite.All, DeviceManagementServiceConfig.ReadWrite.All, DeviceManagementManagedDevices.ReadWrite.All" 
-    }
+    #Assert MS-Graph connection
+    Assert-GraphConnection -Cmdlet $PSCmdlet
 
     # Get all App Protection Policies
     $appProtectionPolicies = Get-ChildItem -Path "$path\App Protection Policies" -File -ErrorAction SilentlyContinue